Task3

+++ Tyuiu.KosyanovDV.Sprint5.Task3.V4.Lib/DataService.cs
@@ -14,7 +14,7 @@ namespace Tyuiu.KosyanovDV.Sprint5.Task3.V4.Lib
 
             if (fileExist) { File.Delete(path); }
 
-            double res = Math.Round(Math.Log((x + 1) / (double)(x + 2)), 2);
+            double res = Math.Round(Math.Log((double)(x + 1) / (double)(x + 2)), 2);
 
             using (BinaryWriter writer = new BinaryWriter(File.Open(path, FileMode.OpenOrCreate), Encoding.Unicode))
             {
